Use Handle common class for Actuator and Joint handle
