+++ hardware_interface/include/hardware_interface/joint_handle.hpp
@@ -17,59 +17,23 @@
 
 #include <string>
 
+#include "hardware_interface/handle.hpp"
 #include "hardware_interface/macros.hpp"
 #include "hardware_interface/visibility_control.h"
 
 namespace hardware_interface
 {
-/** A handle used to get and set the command of a joint on a given interface. */
-class JointHandle
+/** A handle used to get and set a value on a given joint interface. */
+class JointHandle : public Handle<JointHandle>
 {
 public:
   HARDWARE_INTERFACE_PUBLIC
   JointHandle(
     const std::string & name, const std::string & interface_name,
     double * value_ptr = nullptr)
-  : name_(name), interface_name_(interface_name), value_ptr_(value_ptr)
+  : Handle(name, interface_name, value_ptr)
   {
   }
-
-  HARDWARE_INTERFACE_PUBLIC
-  JointHandle with_value_ptr(double * value_ptr)
-  {
-    return JointHandle(name_, interface_name_, value_ptr);
-  }
-
-  HARDWARE_INTERFACE_PUBLIC
-  const std::string & get_name() const
-  {
-    return name_;
-  }
-
-  HARDWARE_INTERFACE_PUBLIC
-  const std::string & get_interface_name() const
-  {
-    return interface_name_;
-  }
-
-  HARDWARE_INTERFACE_PUBLIC
-  double get_value() const
-  {
-    THROW_ON_NULLPTR(value_ptr_);
-    return *value_ptr_;
-  }
-
-  HARDWARE_INTERFACE_PUBLIC
-  void set_value(double value)
-  {
-    THROW_ON_NULLPTR(value_ptr_);
-    *value_ptr_ = value;
-  }
-
-protected:
-  std::string name_;
-  std::string interface_name_;
-  double * value_ptr_;
 };
 
 }  // namespace hardware_interface
